usefull CRUD restFULL Service imlemented

+++ src/main/java/de/elmasconsulting/ecmanagement/Exceptions/UserNotFoundException.java
@@ -2,7 +2,7 @@ package de.elmasconsulting.ecmanagement.Exceptions;
 
 public class UserNotFoundException extends RuntimeException {
 
-    public UserNotFoundException(int id) {
+    public UserNotFoundException(Long id) {
         super("Could not find user " + id);
     }
 }
